[ghstack-poisoned]

+++ torch/sparse/_triton_ops.py
@@ -334,7 +334,7 @@ if _has_triton():
                     mask=mask_k[:, None], other=0.0
                 )
 
-                acc_block += tl.dot(mat1_block, mat2_block, allow_tf32=allow_tf32)
+                acc_block += tl.dot(mat1_block, mat2_block, allow_tf32=allow_tf32, out_dtype=acc_dtype)
 
             if IS_BETA_ZERO:
                 acc_block *= alpha
